requirements.txt
biopython==1.82
certifi==2022.12.7
charset-normalizer==2.1.1
fair-esm==2.0.0
filelock==3.9.0
fsspec==2023.4.0
idna==3.4
iterative-stratification==0.1.7
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
obonet==1.0.0
pandas==2.1.4
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.4.0
tzdata==2023.4
urllib3==1.26.13
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121